refactor: migrate hooks (use-file-size) from web/app/ce to web/app/core
